feat: add signing
